Refactor profile-specific modules to async_upnp_client.profiles
