+++ async_upnp_client/profiles/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Profiles for upnp_client."""
