+++ async_upnp_client/profiles/igd.py
@@ -6,7 +6,7 @@ from ipaddress import IPv4Address
 import logging
 from typing import List, NamedTuple, Optional
 
-from async_upnp_client.profile import UpnpProfileDevice
+from async_upnp_client.profiles.profile import UpnpProfileDevice
 
 
 _LOGGER = logging.getLogger(__name__)
