+++ tests/profiles/__init__.py
@@ -1,0 +1,2 @@
+# -*- coding: utf-8 -*-
+"""Unit tests for profiles."""
