+++ tox.ini
@@ -2,7 +2,7 @@
 envlist = py35, py36, py37, flake8, pylint, pydocstyle, typing
 
 [testenv]
-commands = py.test --cov={envsitepackagesdir}/async_upnp_client --cov-report=term
+commands = py.test --cov=async_upnp_client --cov-report=term
 deps =
     pytest
     pytest-asyncio
